node/requirements.txt
flask>=2.0.0
pyjwt>=2.3.0
sqlite3
python-dotenv>=0.19.0
cryptography>=3.4.0 